requirements.txt
fastapi
python-dotenv
uvicorn[standard]
pandas